feat: add solutions to lc problem: No.2670
No.2670.Find the Distinct Difference Array
