+++ solution/2600-2699/2670.Find the Distinct Difference Array/Solution.py
@@ -1,9 +1,14 @@
 class Solution:
     def distinctDifferenceArray(self, nums: List[int]) -> List[int]:
         n = len(nums)
+        suf = [0] * (n + 1)
+        s = set()
+        for i in range(n - 1, -1, -1):
+            s.add(nums[i])
+            suf[i] = len(s)
+        s.clear()
         ans = [0] * n
-        for i in range(n):
-            a = len(set(nums[: i + 1]))
-            b = len(set(nums[i + 1 :]))
-            ans[i] = a - b
+        for i, x in enumerate(nums):
+            s.add(x)
+            ans[i] = len(s) - suf[i + 1]
         return ans
